fix(release): address release review feedback
